feat(gui_tests): Generate single method URL
In this commit we got a new method, that generates single URL for further GUI testing. Wen want to check if the python extension works, thus we making a single URL address and then we need to send the conformation to the user, that it worked.

BREAKING CHANGE: Added generate single url method