feat(gui_tests, main.py, app.js): Inut wrapper
With this commit we got a wrapper method in main.py that is like a glue between app.js and GUI_tests.py. The data will be sent from JS to Python. In the next commit we should take care of the return value and check if the method works correctly.

BREAKING CHANGE: Added input wrapper from JS into Python