feat(gui_tests.py, app.js, styles.css): Image display
With this commit, we're able to see an image in the display box on the page. It was made using base64. The previous problem was resolved. It was only a one keyword in the frontend, that was not there. The display box is now also a little bigger, so that we can see the whole image without scrolling. Commented code is also removed.

BREAKING CHANGE: Added Base64 image display
Removed commented code