chore(files): Module creation
In this commit we're adding modules to our project, that were made from the app.py. The file was to big, and we wanted that it is easier to maintain and to read. So now we got total 5 new files. app.py is now refactored and works with all the modules. We kept the old app.py file with the name app_old.py for the time of the maintenance. When everything is stable, we're going to delete this file.

BREAKING CHANGE: Added Modules