chore(vizualization.py): New TODO
We need to change the import file to SVG, to keep the resolution on zooming into graph on the html page. When it is not clearly visible tough, we need to convert it with Plotly lib and export it to html. Then we can use it as embed on the html page.