feat(app.js, app.py, main.py, requirements.txt, visualization.py): New plot tech
With this commit, we're changing the technology behind the graphs. Now it is Plotly instead of Matplotlib.
The reason were, that Plotly can display graphs in a interactive manner and save it to .html, which can be embedded in the software.
The has been cleaned and a new requirement has been added.

BREAKING CHANGE: Replaced Matplotlib with Plotly